chore: move util functions into separate file
